pyproject.toml
[project]
name = "django-start-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-crispy-forms>=2.3",
    "django>=5.1.4",
    "crispy-bootstrap5>=2024.10",
    "django-htmx>=1.21.0",
    "htmx-messages",
    "django-extensions>=3.2.3",
    "django-browser-reload>=1.17.0",
    "fontawesomefree==6.6.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=5.0.1",
    "djlint>=1.36.4",
]

[tool.uv.sources]
htmx-messages = { git = "https://github.com/johnnynenha/django_htmx_messages.git" }
